# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tahr Desk",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/agents": {
      "get": {
        "tags": ["admin"],
        "summary": "List Agents",
        "operationId": "admin-list_agents",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersPublic"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create Agent",
        "operationId": "admin-create_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/agents/{user_id}": {
      "patch": {
        "tags": ["admin"],
        "summary": "Update Agent",
        "operationId": "admin-update_agent",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/agents/{user_id}/reset-password": {
      "post": {
        "tags": ["admin"],
        "summary": "Reset Agent Password",
        "operationId": "admin-reset_agent_password",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentResetPassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/site": {
      "get": {
        "tags": ["admin"],
        "summary": "Get Site",
        "operationId": "admin-get_site",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteSettingsPublic"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "tags": ["admin"],
        "summary": "Update Site",
        "operationId": "admin-update_site",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteSettingsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/services": {
      "get": {
        "tags": ["admin"],
        "summary": "List Services",
        "operationId": "admin-list_services",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicesAdmin"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create Service",
        "operationId": "admin-create_service",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCategoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceCategoryPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/services/{service_id}": {
      "patch": {
        "tags": ["admin"],
        "summary": "Update Service",
        "operationId": "admin-update_service",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Service Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCategoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceCategoryPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/webhooks": {
      "get": {
        "tags": ["admin"],
        "summary": "List Webhooks",
        "operationId": "admin-list_webhooks",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhooksPublic"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create Webhook",
        "operationId": "admin-create_webhook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/webhooks/{endpoint_id}": {
      "patch": {
        "tags": ["admin"],
        "summary": "Update Webhook",
        "operationId": "admin-update_webhook",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Endpoint Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/webhooks/{endpoint_id}/rotate-secret": {
      "post": {
        "tags": ["admin"],
        "summary": "Rotate Webhook Secret",
        "operationId": "admin-rotate_webhook_secret",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Endpoint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/webhooks/{endpoint_id}/test": {
      "post": {
        "tags": ["admin"],
        "summary": "Test Webhook",
        "operationId": "admin-test_webhook",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Endpoint Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Full Name"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": ["email", "full_name", "password"],
        "title": "AgentCreate"
      },
      "AgentResetPassword": {
        "properties": {
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": ["new_password"],
        "title": "AgentResetPassword"
      },
      "AgentUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "new_password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 12
              },
              {
                "type": "null"
              }
            ],
            "title": "New Password"
          }
        },
        "type": "object",
        "title": "AgentUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": ["message"],
        "title": "Message"
      },
...
      "ServiceCategoryCreate": {
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "summary": {
            "type": "string",
            "maxLength": 500,
            "minLength": 10,
            "title": "Summary"
          },
          "guidance_markdown": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Guidance Markdown"
          },
          "response_target_hours": {
            "type": "integer",
            "maximum": 8760.0,
            "minimum": 1.0,
            "title": "Response Target Hours"
          },
          "sort_order": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Sort Order",
            "default": 0
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "summary",
          "guidance_markdown",
          "response_target_hours"
        ],
        "title": "ServiceCategoryCreate"
      },
      "ServiceCategoryPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "guidance_html": {
            "type": "string",
            "title": "Guidance Html"
          },
          "response_target_hours": {
            "type": "integer",
            "title": "Response Target Hours"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "summary",
          "guidance_html",
          "response_target_hours",
          "sort_order",
          "is_active"
        ],
        "title": "ServiceCategoryPublic"
      },
...
      "ServicesAdmin": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ServiceCategoryAdmin"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "ServicesAdmin"
      },
...
      "SiteSettingsPublic": {
        "properties": {
          "organization_name": {
            "type": "string",
            "title": "Organization Name"
          },
          "service_area": {
            "type": "string",
            "title": "Service Area"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "introduction_html": {
            "type": "string",
            "title": "Introduction Html"
          },
          "webhooks_enabled": {
            "type": "boolean",
            "title": "Webhooks Enabled"
          },
          "introduction_markdown": {
            "type": "string",
            "title": "Introduction Markdown"
          }
        },
        "type": "object",
        "required": [
          "organization_name",
          "service_area",
          "timezone",
          "introduction_html",
          "webhooks_enabled",
          "introduction_markdown"
        ],
        "title": "SiteSettingsPublic"
      },
      "SiteSettingsUpdate": {
        "properties": {
          "organization_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Organization Name"
          },
          "service_area": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Service Area"
          },
          "timezone": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone"
          },
          "introduction_markdown": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Introduction Markdown"
          }
        },
        "type": "object",
        "required": [
          "organization_name",
          "service_area",
          "timezone",
          "introduction_markdown"
        ],
        "title": "SiteSettingsUpdate"
      },
...
      "UserPublic": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "is_active",
          "is_superuser",
          "full_name",
          "id",
          "role",
          "created_at"
        ],
        "title": "UserPublic"
      },
      "UsersPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/UserPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "UsersPublic"
      },
...
      "WebhooksPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/WebhookPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "WebhooksPublic"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}